dwc2: Fix USB send/receive to unaligned memory addresses
Here is a very small patch for what I think is a rather severe
issue with the dwc2 driver which is not 64 bit specific. It looks
like the bug was a simple typo, but it means that USB
send/receive to unaligned memory addresses is broken in the stock driver.
It may or may not be necessary to apply this fix.

Original author: Electron752
https://www.raspberrypi.org/forums/viewtopic.php?p=945129#p945129