bcm2710: fix dma control block data structure
DMA controller is fed with a linked list of control block structures
which describe the data transfers to be done. The fields of these
structs are unsigned 32 bits integers.

Unfortunately the kernel driver assumes that an unsigned long is
always a 32 bits integer. This assumption is correct for armv7 arch, but
it fails for armv8 where the unsigned long is a 64 bits integer.

This fix replaces the "unsigned long" type with the more portable
"uint32_t"